refactor: 사용하지 않는 hook을 제거하고 error validation로직을 custom hook으로 분리한다
