--- src/context/Form/FormChangeContext.tsx
@@ -1,59 +1,0 @@
-import { createContext, Dispatch, ReactNode, useReducer } from 'react'
-import { CardBaseFormProps } from '../../components/Form'
-
-type PropName = keyof CardBaseFormProps
-type Payload = { name: PropName }
-type Action = { type: 'CHANGE'; payload: Payload }
-type FormChangedContextProps = Record<PropName, boolean>
-type FormDispatcher = Dispatch<Action>
-
-export const formChangedContext = createContext<
-  FormChangedContextProps | undefined
->(undefined)
-export const formChangeedDispatcherContext = createContext<
-  FormDispatcher | undefined
->(undefined)
-
-function getInitialValue() {
-  return {
-    cvc: false,
-    number1: false,
-    number2: false,
-    number3: false,
-    number4: false,
-    expireAtYear: false,
-    expireAtMonth: false,
-    owner: false,
-    password1: false,
-    password2: false,
-    type: false,
-    name: false,
-  }
-}
-function formReducer(
-  state: FormChangedContextProps,
-  action: Action
-): FormChangedContextProps {
-  switch (action.type) {
-    case 'CHANGE':
-      const newState = getInitialValue()
-      newState[action.payload.name] = true
-      return newState
-  }
-}
-
-export default function FormContextProvider({
-  children,
-}: {
-  children: ReactNode
-}) {
-  const [state, dispatch] = useReducer(formReducer, getInitialValue())
-
-  return (
-    <formChangeedDispatcherContext.Provider value={dispatch}>
-      <formChangedContext.Provider value={state}>
-        {children}
-      </formChangedContext.Provider>
-    </formChangeedDispatcherContext.Provider>
-  )
-}
