+++ src/context/Form/hooks/index.ts
@@ -1,21 +1,5 @@
 import { useContext } from 'react'
 import { cardFormContext, cardFormDispatcherContext } from '../CardFormContext'
-import {
-  formChangedContext,
-  formChangeedDispatcherContext,
-} from '../FormChangeContext'
-
-function useFormChangedState() {
-  const state = useContext(formChangedContext)
-  if (!state) throw new Error('State Not Found!')
-  return state
-}
-
-function useFormChangedDispatch() {
-  const dispatch = useContext(formChangeedDispatcherContext)
-  if (!dispatch) throw new Error('Dispatcher Not Founct!')
-  return dispatch
-}
 
 function useCardFormState() {
   const state = useContext(cardFormContext)
@@ -29,9 +13,4 @@ function useCardFormDispatch() {
   return dispatch
 }
 
-export {
-  useFormChangedDispatch,
-  useFormChangedState,
-  useCardFormState,
-  useCardFormDispatch,
-}
+export { useCardFormState, useCardFormDispatch }
